wip: adapt consumer / producer so the number of thread is not fixed by the partition, TU ok for chronicles

+++ nuxeo-mqueues-core/src/main/java/org/nuxeo/ecm/platform/importer/mqueues/pattern/producer/ProducerPool.java
@@ -39,7 +39,7 @@ public class ProducerPool<M extends Message> extends AbstractCallablePool<Produc
     private final ProducerFactory<M> factory;
     private final String mqName;
 
-    public ProducerPool(String mqName, MQManager<M> manager, ProducerFactory<M> factory, int nbThreads) {
+    public ProducerPool(String mqName, MQManager<M> manager, ProducerFactory<M> factory, short nbThreads) {
         super(nbThreads);
         this.mqName = mqName;
         this.manager = manager;
